chore: extract admin flags handler (#298 5b)
Move GET/POST /admin/merch/flags business logic out of
ingest/routes/admin-flags.ts into ingest/admin-flags-handler.ts.

Routes file now only does auth allowlist + logOutcome + delegate,
matching the handler/routes split used by likes/bookmarks/follows
and admin-orders. No behavior change — same flagsStore handling,
same status/header/error_code shapes, same 500/400/200 paths.

Note: ingest/admin-handler.ts still contains duplicate
handleGetMerchFlags/handleSetMerchFlags from earlier iteration;
follow-up can either delete them or make them re-export from
the new handler to keep a single source. Left as-is to keep
this split minimal.

Verified: tsc pass, prettier pass, 854 tests pass.
